add caching via .htaccess redirects to cached files
this is probably as fast as it gets with only techniques that will be available
on virtually every wordpress installation; only rendering all the static files
right on save_post instead of waiting for the first request will still be
faster (and only faster on the first request).

this feature is not production ready! there is no way to invalidate the cache
right now! also the api for switching the cache on will likely have to change!
and probably the implementation is buggy! no idea if sha256 is usually available
on php5 hosts!