CAUTION API CHANGE add cache invalidation on post_save
The unplug\prevent_parse_request function is now simply called unplug\unplug and
should be used in functions.php to init all kinds of unplug behaviour. By
default, it still just prevents WordPress from handling certain requests so that
an unplug\Router instance may be used in index.php instead. But it now also
takes and optional $options array that is currently used to switch on caching
and pass an individual cache dir. Also, if caching is on, cache will be flushed
on post_save now. And the caching-related parameters from the unplug\Router
constructor are removed.